refactor(article,draft): integrate draft application to article application
