+++ services/article/application/src/main/java/nettee/blolet/article/application/port/DraftBlockQueryPort.java
@@ -1,14 +1,11 @@
-package nettee.draft.draftblock.application.port;
+package nettee.blolet.article.application.port;
 
-import nettee.draft.draftblock.domain.type.DraftBlockStatus;
+import nettee.blolet.article.domain.type.DraftBlockStatus;
 import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockDetail;
 import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockSummary;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 public interface DraftBlockQueryPort {
     Optional<DraftBlockDetail> findById(String id);
