+++ services/article/application/src/main/java/nettee/blolet/article/application/port/DraftCommandPort.java
@@ -1,9 +1,9 @@
-package nettee.draft.application.port;
+package nettee.blolet.article.application.port;
 
 import nettee.draft.domain.DraftImage;
 import nettee.draft.readmodel.DraftReadModels.DraftDetail;
 import nettee.draft.domain.Draft;
-import nettee.draft.domain.type.DraftStatus;
+import nettee.blolet.article.domain.type.DraftStatus;
 import java.util.Optional;
 
 public interface DraftCommandPort {
