+++ services/article/application/src/main/java/nettee/blolet/article/application/service/DraftQueryService.java
@@ -1,15 +1,15 @@
-package nettee.draft.application.service;
+package nettee.blolet.article.application.service;
 
 import lombok.RequiredArgsConstructor;
-import nettee.blolet.draft.api.validation.context.DraftContextValidationSupplier;
+import nettee.blolet.article.api.validation.context.DraftContextValidationSupplier;
 import nettee.common.validation.model.ValidationResponseModel;
-import nettee.draft.application.usecase.DraftValidationResponseUseCase;
+import nettee.blolet.article.application.usecase.DraftValidationResponseUseCase;
 import nettee.draft.readmodel.DraftReadModels.DraftDetail;
 import nettee.draft.readmodel.DraftReadModels.DraftSummary;
-import nettee.draft.application.port.DraftQueryPort;
-import nettee.draft.domain.type.DraftStatus;
-import nettee.draft.application.usecase.DraftReadByStatusesUseCase;
-import nettee.draft.application.usecase.DraftReadUseCase;
+import nettee.blolet.article.application.port.DraftQueryPort;
+import nettee.blolet.article.domain.type.DraftStatus;
+import nettee.blolet.article.application.usecase.DraftReadByStatusesUseCase;
+import nettee.blolet.article.application.usecase.DraftReadUseCase;
 import nettee.draft.readmodel.DraftReadModels.DraftTitle;
 import org.springframework.stereotype.Service;
