+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftBlockDeleteUseCase.java
@@ -1,4 +1,4 @@
-package nettee.draft.draftblock.application.usecase;
+package nettee.blolet.article.application.usecase;
 
 public interface DraftBlockDeleteUseCase {
     void delete(String userId, String draftBlockId);
