+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftBlockUpdateUseCase.java
@@ -1,4 +1,4 @@
-package nettee.draft.draftblock.application.usecase;
+package nettee.blolet.article.application.usecase;
 
 import nettee.draft.draftblock.domain.DraftBlock;
