+++ services/article/application/src/main/java/nettee/blolet/article/application/usecase/DraftCreateUseCase.java
@@ -1,4 +1,4 @@
-package nettee.draft.application.usecase;
+package nettee.blolet.article.application.usecase;
 
 import nettee.draft.domain.Draft;
